queue: split struct AvailIter into submodule
Split `struct AvialIter` into a submodule.
